Update of last snowballing commit

+++ BasicBrowser/upload_docs.py
@@ -143,10 +143,10 @@ def main():
     q.r_count = n_records
     q.save()
 
-
-    shutil.rmtree("/queries/"+title)
-    os.remove("/queries/"+title+".txt")
-    os.remove("/queries/"+title+".log")
+    if q.type == "default":
+        shutil.rmtree("/queries/"+title)
+        os.remove("/queries/"+title+".txt")
+        os.remove("/queries/"+title+".log")
 
 
 if __name__ == '__main__':
